Use Bob swap factory in production
