+++ swap/src/protocol/bob.rs
@@ -1,15 +1,16 @@
 //! Run an XMR/BTC swap in the role of Bob.
 //! Bob holds BTC and wishes receive XMR.
+use anyhow::Result;
 use libp2p::{core::Multiaddr, NetworkBehaviour, PeerId};
 use tracing::{debug, info};
 
 use crate::{
     bitcoin,
     bitcoin::EncryptedSignature,
-    monero,
+    database, monero, network,
     network::peer_tracker::{self, PeerTracker},
     protocol::{alice, bob},
-    SwapAmounts,
+    StartingBalances, SwapAmounts,
 };
 
 pub use self::{
@@ -22,8 +23,10 @@ pub use self::{
     state::*,
     swap::{run, run_until},
 };
-use crate::database::Database;
-use std::sync::Arc;
+use crate::{config::Config, database::Database, network::transport::build, seed::Seed};
+use libp2p::identity::Keypair;
+use rand::rngs::OsRng;
+use std::{path::PathBuf, sync::Arc};
 use uuid::Uuid;
 
 mod amounts;
@@ -42,6 +45,160 @@ pub struct Swap {
     pub bitcoin_wallet: Arc<bitcoin::Wallet>,
     pub monero_wallet: Arc<monero::Wallet>,
     pub swap_id: Uuid,
+}
+
+pub struct BobSwapFactory {
+    identity: Keypair,
+    peer_id: PeerId,
+
+    db_path: PathBuf,
+    swap_id: Uuid,
+
+    pub bitcoin_wallet: Arc<bitcoin::Wallet>,
+    pub monero_wallet: Arc<monero::Wallet>,
+    config: Config,
+    pub starting_balances: StartingBalances,
+
+    alice_connect_address: Multiaddr,
+    alice_connect_peer_id: PeerId,
+}
+
+impl BobSwapFactory {
+    #[allow(clippy::too_many_arguments)]
+    pub fn new(
+        seed: Seed,
+        db_path: PathBuf,
+        swap_id: Uuid,
+        bitcoin_wallet: Arc<bitcoin::Wallet>,
+        monero_wallet: Arc<monero::Wallet>,
+        config: Config,
+        starting_balances: StartingBalances,
+        alice_connect_address: Multiaddr,
+        alice_connect_peer_id: PeerId,
+    ) -> Self {
+        let identity = network::Seed::new(seed).derive_libp2p_identity();
+        let peer_id = identity.public().into_peer_id();
+
+        Self {
+            identity,
+            peer_id,
+            db_path,
+            swap_id,
+            bitcoin_wallet,
+            monero_wallet,
+            config,
+            starting_balances,
+            alice_connect_address,
+            alice_connect_peer_id,
+        }
+    }
+
+    pub async fn new_swap_as_bob(
+        &self,
+        swap_amounts: SwapAmounts,
+    ) -> Result<(bob::Swap, bob::EventLoop)> {
+        let initial_state = init_bob_state(
+            swap_amounts.btc,
+            swap_amounts.xmr,
+            self.bitcoin_wallet.clone(),
+            self.config,
+        )
+        .await?;
+
+        let (event_loop, event_loop_handle) = init_bob_event_loop(
+            self.identity.clone(),
+            self.peer_id.clone(),
+            self.alice_connect_peer_id.clone(),
+            self.alice_connect_address.clone(),
+        )?;
+
+        let db = Database::open(self.db_path.as_path())?;
+
+        Ok((
+            Swap {
+                state: initial_state,
+                event_loop_handle,
+                db,
+                bitcoin_wallet: self.bitcoin_wallet.clone(),
+                monero_wallet: self.monero_wallet.clone(),
+                swap_id: self.swap_id,
+            },
+            event_loop,
+        ))
+    }
+
+    pub async fn recover_bob_from_db(&self) -> Result<(bob::Swap, bob::EventLoop)> {
+        // reopen the existing database
+        let db = Database::open(self.db_path.clone().as_path())?;
+
+        let resume_state = if let database::Swap::Bob(state) = db.get_state(self.swap_id)? {
+            state.into()
+        } else {
+            unreachable!()
+        };
+
+        let (event_loop, event_loop_handle) = init_bob_event_loop(
+            self.identity.clone(),
+            self.peer_id.clone(),
+            self.alice_connect_peer_id.clone(),
+            self.alice_connect_address.clone(),
+        )?;
+
+        Ok((
+            Swap {
+                state: resume_state,
+                event_loop_handle,
+                db,
+                bitcoin_wallet: self.bitcoin_wallet.clone(),
+                monero_wallet: self.monero_wallet.clone(),
+                swap_id: self.swap_id,
+            },
+            event_loop,
+        ))
+    }
+}
+
+async fn init_bob_state(
+    btc_to_swap: bitcoin::Amount,
+    xmr_to_swap: monero::Amount,
+    bob_btc_wallet: Arc<bitcoin::Wallet>,
+    config: Config,
+) -> Result<BobState> {
+    let amounts = SwapAmounts {
+        btc: btc_to_swap,
+        xmr: xmr_to_swap,
+    };
+
+    let refund_address = bob_btc_wallet.new_address().await?;
+    let state0 = bob::State0::new(
+        &mut OsRng,
+        btc_to_swap,
+        xmr_to_swap,
+        config.bitcoin_cancel_timelock,
+        config.bitcoin_punish_timelock,
+        refund_address,
+        config.monero_finality_confirmations,
+    );
+
+    Ok(BobState::Started { state0, amounts })
+}
+
+fn init_bob_event_loop(
+    identity: Keypair,
+    peer_id: PeerId,
+    alice_peer_id: PeerId,
+    alice_addr: Multiaddr,
+) -> Result<(bob::event_loop::EventLoop, bob::event_loop::EventLoopHandle)> {
+    let bob_behaviour = bob::Behaviour::default();
+    let bob_transport = build(identity)?;
+
+    bob::event_loop::EventLoop::new(
+        bob_transport,
+        bob_behaviour,
+        peer_id,
+        alice_peer_id,
+        alice_addr,
+    )
 }
 
 #[derive(Debug, Clone)]
